refactor(utilities): integrate siege_utilities, move GST-specific helpers to locations/ (#28)
Closes #28. Rips out app/hellodjango/utilities/ (proto-siege_utilities)
and replaces with siege_utilities for parity-cases, locations/helpers/
for GST-unique helpers, and locations/dispatchers.py + locations/data/
for constants and data fixtures.

What moves where:
- siege_utilities.files.remote.download_file (was utilities/file_utilities.py)
- siege_utilities.files.paths.ensure_path_exists (was utilities/file_utilities.py)
- siege_utilities.files.hashing.calculate_file_hash (was utilities/file_utilities.py
  generate_sha256_hash_for_file AND utilities/dataset_cache.py calculate_file_hash)
- locations/helpers/file_ops.py: run_subprocess, generate_local_path_from_url,
  unzip_file_to_its_own_directory, add_hash_entry_to_dispatcher,
  check_for_hash_in_dispatcher (GST-unique behaviors)
- locations/helpers/geocoding.py: distance_to_decimal_degrees, geocode_with_nominatim_public
- locations/helpers/spatial_data.py: find_vector_dataset_file_in_directory,
  fix_gadm_null_foreign_keys, GADM_MODEL_FIELD_NAMES
- locations/helpers/django_admin.py: get_random_object, create_united_states_address,
  update_model_fields_to_single_value, update_model_geometry_foreign_keys
- locations/helpers/dataset_cache.py: cache-manifest read/write (calculate_file_hash
  imported from SU)
- locations/dispatchers.py: DOWNLOADS_DISPATCHER + all dispatcher constants
- locations/data/file_hashes.py: FILE_NAMES_AND_HASHES

Dropped: reset_primary_keys (empty stub, 0 callers).
Removed dead: views/interactive.py had `from utilities import *` with 0 usage.

Adds siege-utilities[geo]>=3.17.2 to docker/requirements.txt.

The GST-unique helpers in locations/helpers/ are candidates for porting
up to siege_utilities (tracked at siege-analytics/siege_utilities#515).
When that lands, a follow-up GST PR can shrink locations/helpers/ to
genuinely-GST-specific code only.

Verified in container with full GST requirements installed:
- python manage.py check: 0 issues
- python manage.py makemigrations --check --dry-run: No changes detected

Self-Review: utilities/ rip-out clean; all 5 wildcards replaced with explicit imports; container-validated; SU follow-up filed
Self-Review-Source: /Users/dheerajchand/.craft-agent/workspaces/my-workspace/sessions/260502-vital-channel/plans/review-gst-28-utilities-refactor.md
