+++ app/hellodjango/locations/data/file_hashes.py
