--- app/hellodjango/utilities/__init__.py
@@ -1,6 +1,0 @@
-from .file_utilities import *
-from .dispatchers import *
-from .vector_data_utilities import *
-from .existing_file_hashes import *
-from .django_model_management import *
-from .geocoding import *
