+++ docker/requirements.txt
@@ -2,6 +2,7 @@ black
 boto3
 Django
 django-cors-headers
+siege-utilities[geo]>=3.17.2
 django-extensions
 django-grappelli
 djangorestframework
